Token moves now based on TokenId matching tile instead of tokenposition
Tokens actually carry their tiles with them so we need to check
if a tokens ID matches the tile under it to know if it wants to move.
The token always matches the tile it sits on, so if we are moving
then the token's id will not match the tile to which we are moving.